PEEL Shopping 7.0.3
- Ajout de robustesse sur sélection automatique de langue du site (utile
lors de MAJ de site via FTP : si incohérences entre fichiers on pouvait
se retrouver dans cas où session_langue est vide)
- Ajout de variables de configurations
- Ajout de is_lot_module_active manquant avant l'appel de la table
peel_quantites.
- Ajout de la fonction de conversion de tag dans le contenu des
articles.
- Ajout de placeholder http:// sur url pour les input spécifiques url
dans les formulaires
- Ajout des promotions par catégorie par seuil de produits commandé.
- Ajout de l'appel à Google Analytics sur pages non HTML : flux RSS, et
flux pour application iphone si module installé
- Ajout de overflow:hidden dans .slide => les carrousels, quand ils se
chargent, ne déborderont plus en dessous du bloc
- Amélioration de la recherche de mots clés dans un texte
- Amélioration de redirections pour référencement
- Amélioration de la conformité HTML 5
- Amélioration de la détection de TVA intracom pour exonérations ou non
- Amélioration de la gestion bannières flash sur iOS : possibilité de
mettre alternative dans tag_html de peel_banniere
- Amélioration de la gestion des metas : possibilité de remplir les
données par URL dans l'administration
- Amélioration de la gestion pour convertir des points en cadeaux
- Amélioration de la gestion thumbs si image pas valide
- Améliorations diverses pour le référencement
- Modification des inputs => attributs HTML5 pour les input text :
email, tel, search et number - on n'utilise pas les autres car on a déjà
datepicker en HTML5
- Correction arguments de setcookie si site en multidomaines
- Correction CSS lightbox (si on l'active à la place de jqzoom)
- Correction d'un appel à une constante de langue
- Correction d'un message de confirmation dans l'admin.
- Correction de l'affichage de remise par lot dans les pages de
catégories
- Correction de la gestion multilingue des carrousels
- Correction inscription si on a déjà un compte 'newsletter'
- Correction textes, notamment explication du mot secret Moneybookers
- Correction version minimale PHP : 5.1.2
- Création de paramétrage pour gestion des champs spécifiques et des
champs obligatoires pour formulaire d'inscription.
- Dorénavant les pubs .swf ne sont plus envoyées aux ipad/iphone/ipod si
pub flash seule dans container de pub (si avec pub html image en + dans
son espace, alors on envoie quand même pour avoir la pub image)
- Eviter l'envoi d'un email de template vide (on ne rajoute la signature
que si pas vide)
- Facture à partir de l'historique utilisateur des commandes uniquement
si en payé
- Interconnexion avec analytics pour pages qui ne sont pas en HTML +
filtrage des moteurs pour appel à analytics pour pages RSS et
application iphone
- Gestion des langues administrables mais pas visibles en front-office :
nouveau statuts, avec puce orange
- Gestion du prix additionnel d'un attribut si option unique possible =>
on intègre automatiquement le prix de l'attribut au produit
- Les anciens navigateurs traitent les valeurs type inconnues comme text
=> pas de problème de compatibilité
- Limitation du nombre de mots clés dans meta keywords
- Modification de get_menu pour permettre d'administrer avec variables
de configuration les sous menus.
- Modification de la source du champ gold_credit
- Optimisation du code
- Suppression de paramétrage présent directement dans info.inc.php =>
mis dans table de configuration
- Textes moins coupés dans liste de catégories
- Utilisation de la fonction is_gifts_module_active update_points
(déplacée dans le module gift) + Corrections de la fonctionnalité de
produits cadeaux.

+++ modules/pensebete/voir.php
@@ -3,14 +3,14 @@
 // +----------------------------------------------------------------------+
 // | Copyright (c) 2004-2013 Advisto SAS, service PEEL - contact@peel.fr  |
 // +----------------------------------------------------------------------+
-// | This file is part of PEEL Shopping 7.0.2, which is subject to an	  |
+// | This file is part of PEEL Shopping 7.0.3, which is subject to an	  |
 // | opensource GPL license: you are allowed to customize the code		  |
 // | for your own needs, but must keep your changes under GPL			  |
 // | More information: https://www.peel.fr/lire/licence-gpl-70.html		  |
 // +----------------------------------------------------------------------+
 // | Author: Advisto SAS, RCS 479 205 452, France, https://www.peel.fr/	  |
 // +----------------------------------------------------------------------+
-// $Id: voir.php 36232 2013-04-05 13:16:01Z gboussin $
+// $Id: voir.php 36927 2013-05-23 16:15:39Z gboussin $
 
 include("../../configuration.inc.php");
 
@@ -25,7 +25,7 @@ define('IN_PENSE_BETE', true);
 if(! empty($_GET['mode'])){
 	if($_GET['mode']== 'delete'){
 		if(!empty($_GET['id'])){
-			$sql_delete= 'DELETE FROM peel_pensebete WHERE id='.intval($_GET['id']);
+			$sql_delete= 'DELETE FROM peel_pensebete WHERE id='.intval($_GET['id']). (!a_priv('admin*', false)?' AND id_utilisateur="' . intval($_SESSION['session_utilisateur']['id_utilisateur']) . '"':'');
 			query($sql_delete);
 		}
 	}
@@ -34,9 +34,9 @@ if(! empty($_GET['mode'])){
 include($GLOBALS['repertoire_modele'] . "/haut.php");
 if(is_annonce_module_active()) {
 	display_ads_in_reminder();
-} else {
-	display_product_in_reminder();
+	echo '<br />';
 }
+display_product_in_reminder();
 include($GLOBALS['repertoire_modele'] . "/bas.php");
 
 ?>
